potential fix of skipping weekly cert
weekly cert was skipped after weekly work search entry was completed for some reason and the program quit normally as if all work was done. this might be caused by screenID not getting correct ID because the text contents doesn't load fast enough at the same time as its element.