dynamic relative file paths
packaging the project (moving all the project files into a subfolder) for the pipenv-to-poetry migration means that file finding attempts will get invalid relative paths since all the execution path roots are the upmost project folder, not the package subfolder. all relative paths are now dynamically converted to full system paths so that they are always valid no matter where the files are as long as the provided relative path is itself valid.